Parse args from util, allow arbitrary args
